Refactor useFormData hook to accept generic form data
